Simplified rule and changed example/format
I added a note for the code smell. We don't have a rule to link to
yet regarding too many assignments. That rule seems more related to
OOP rather than just format and styling within Ruby. I have an escaped
asterisk because it was causing issues with VIM syntax highlighting.
Also added a new line to try and stop the error message about missing a
newline at end of file.